Changed keyword relation modification
Not every keyword changes the relation, but unknown ones do, also NPCs
might have different reactions for the standard keywords, this makes an
overrideable GetStandardKeywordResponse very impractical.

I've changed this to a ChangeRelation method, that is called from the
NPCs when needed. This method changes the relation values and sends the
random reaction message. This requires calls to the method in every NPC,
but that's probably better than handling all keywords twice in another
method that has to be overriden for every NPC.

Also, I've changed the mood message to only be displayed at the
beginning of the conversation, this and the way the relation change
works now seems to be the official behavior.