Changed loading of item scripts
First change in the way items are loaded, creating the script file for
the inline scripts in a cleaner way, placing it in cache, loading it
through the script list. This breaks items for the moment. In the future
attributes on the item script classes will dictate what items are
handled by which script, instead of the "use" construct, file, and class
names.