Fixed Defense being able to be hit twice
This actually seemed to have caused a bug where you could get locked
completely with the old CBS enabled. Your Defense would, presumably, be
hit twice, a Run unlock would be sent for the first one, but not the
second, because new unlocks are only sent when they change the
creature's current locks. After this fix I can't reproduce this anymore.
Even though I'm still not 100% sure how it happened.