Hot-fixed cutscene locks
The unlock wasn't sent because the locks didn't change when a skill ran
out during the cutscene. For example, if FH ends during a cutscene all
locks are lifted in SkillCancel. Since the following unlock in the
cutscene doesn't change any locks (they were lifted already) the unlock
wasn't sent. We have to implement counters for the locks.