Fixed chat invitation logic
Previously, when starting a chat, the friend would be put on a waiting
list, until the first msg got in. Then they were moved to the normal
user list. While this design kinda worked, it doesn't properly support
the way official handles sessions. You're supposed to always get back
into the same session, even across relogs. But that's not possible if we
have only a single user list. We have to remove the user on logout, to
get rid of the reference, but then we'd have to "reinsert" the new user,
after relog. Now we have a second list of user's ids that are supposed
to be part of the chat.