Fixed extensions for client props
We didn't send them, because there was not left over int for them in the
client prop packet, but apparently they are sent, if there are any. This
fixes a problem where you would occasionally not get a confirmation
message, because you didn't get the extension from the appears packet.