Fixed left hand auto-move check
When switching equipment on the right hand, whatever is in the left
hand/magazine should be unequipped. The only type of item I know that
does not get unequipped is shields, so we will check for that for now,
getting rid of the very weird previous check.

This also fixes manuals not being unequipped automatically when
unequipping the hammer.