Fixed #74 - Combat action skill id usage
Thsi includes a long overdue refactoring of how combat action skill ids
are used. Originally we didn't understand the purpose of the skill id in
that target actions, so we set it to the attacker skill and used it
freely in all skill training. We now know that skill id must be the id
of the currently active skill of the target, which changes how we handle
training a little. This change was necessary to fix the bug, we checked
for Defense being *active*, which might not be the case, when it's
canceled in the cap handling.