Disabled Artisan upgrades
We don't need them for our current goals, and we lack research data to
implement them, but the client will still display them, and the server
didn't warn the player that the effect won't be applied. Now it doesn't
let you apply it anymore for the time being.
Related to #134.