Moved web registration to an optional script
Unlike the hotkey saving, visual chat, etc, the registration is not an
essential part of the game. Our current registration page is also very
bare bones, and it probably wouldn't even be used in production. It
makes more sense to put it into an optional script, so admins can
activate it when they want (to provide) a simple way to create accounts,
or they can leave it disabled, if they don't need or want it.