Fixed errors in scripting var handling
Invariant wasn't enforced for saving the variables as strings in the db,
so the system culture was used. If you aren't using the US date format,
you would get an error when trying to login, because after the
localization updates, the default is en-US, which isn't compatible to
the format dates were saved as in the db. In case of that exception, the
var reader now tries to read the date with the installed UI culture.