Fixed #320 - Exception during upgrades
While implementing the uprgade effects I had to change the type of the
effect value list to float. I also changed the JSON value type to check
for to float, but values that don't explicitly look like a float are
considered an int, so it didn't add the effect value, the list was
empty, and during an upgrade you would get an out of bounds exception.