Fixed item id duplication during trade
The usual problem, if no new item is created we get a duplicate id in
the database. The client also won't display the new items, since it
still believes them to be in the trade window, based on their id. Though
that second part can probably be fixed by sending TradeItemRemove during
the trade completion.