Bug fix: Do not count WIP items (resubmission)
Collect quest objectives will no longer count tailoring/blacksmithing works in progress.

This was solved by adding checks for whether PRGRATE metadata exists on the item to be counted. The item counting functions were also overloaded to allow the caller to specify their own conditions for how and when an item gets counted.