Moved ActiveSkillWait skill regen
It was in the handler, but if we cancel skills on the server we don't go
through the handler. It needs to be in the general canceler. This caused
skills to keep using Mana/Stamina, even after they were canceled by the
server. In particular, this affected Final Hit, which is now canceled by
the server.