Fixed game event broadcasts
Since the client repeats the event broadcast regularly and it wasn't
updated when an event ended, it kept broadcasting events after they were
over. Now an update is sent when an event is started/over. I also
removed the broadcast on ErinnDayTimeTick. With the client repeating the
message, we don't need another manual broadcast.